pyproject.toml
[tool.poetry]
name = "try-study-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.110.2"}
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
fastapi-cache2 = "^0.2.1"
redis = "^5.0.4"
celery = "^5.4.0"
flower = "^2.0.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "app",
]

asyncio_mode="auto"

